+++ src/components/editor/input_manager/input-manager.tsx
@@ -1,142 +1,11 @@
 import { BaseNode } from "~/components/nodes/base-node";
 import { NodeSlot } from "~/components/nodes/slot/node-slot";
 import { ComponentEventHandler } from "../tools/base-tool";
-import { createSignal } from "solid-js";
 import { createStore, SetStoreFunction } from "solid-js/store";
+import { Keybind, KeyEventData, KeyModifiers, MBUTTON_CODES, MouseButtons } from "./keybind-events";
+import { EventHandler, InputEvents } from "./event-handling";
 
-// TODO: Change this to be an Events enum
-// then create a set_event_handler function on KeyEventManager
-export enum InputEvents {
-    POINTER_MOVING,
-    CLICK_ON_NODE,
-    CLICK_ON_NODE_SLOT,
-    HOVER_NODE,
-    HOVER_SLOT
-}
-
-export enum MouseButtons {
-    NONE = 0,
-    LEFT = 1,
-    RIGHT = 2,
-    MIDDLE = 4,
-    SCROLL
-}
-const MBUTTON_CODES = Object.values(MouseButtons).filter((item): item is number => typeof item === 'number');
-
-export enum KeyModifiers {
-    CTRL,
-    SHIFT,
-    ALT
-}
-
-export interface KeybindMapBinds {
-    mouse_buttons?: MouseButtons[], 
-    keys?: string[], 
-    modifiers?: KeyModifiers[]
-}
-
-export class KeybindMap {
-    private _mouse_buttons: MouseButtons[];
-    private _keys: string[];
-    private _modifiers: KeyModifiers[];
-    private _active: () => boolean;
-    private _set_active: (v: boolean) => void;
-
-    constructor(binds: KeybindMapBinds) {
-        const [active, setActive] = createSignal(false);
-        this._active = active;
-        this._set_active = setActive;
-        
-        this._mouse_buttons = binds.mouse_buttons == null ? [] : binds.mouse_buttons;
-        this._keys = binds.keys == null ? [] : binds.keys;
-        this._modifiers = binds.modifiers == null ? [] : binds.modifiers;
-    }
-
-    get active() { return this._active(); }
-    set active(value: boolean) { this._set_active(value); }
-
-    get mouse_buttons() { return this._mouse_buttons; }
-    get keys() { return this._keys; }
-    get modifiers() { return this._modifiers; }
-
-    public update_active(keys: Record<string, boolean>, mouse_buttons: Record<number, boolean>, modifiers: Record<number, boolean>): boolean {
-        const keys_active = this._keys.length > 0 ? this._keys.every(k => !!keys[k]) : true;
-        const mb_active = this._mouse_buttons.length > 0 ? this._mouse_buttons.every(m => !!mouse_buttons[m]) : true;
-        const mod_active = this._modifiers.length > 0 ? this._modifiers.every(mod => !!modifiers[mod]) : true;
-
-        const has_any_bind = this._keys.length > 0 || this._mouse_buttons.length > 0 || this._modifiers.length > 0;
-        const is_now_active = has_any_bind && keys_active && mb_active && mod_active;
-
-        if (this.active !== is_now_active) {
-            this.active = is_now_active;
-        }
-        return is_now_active;
-    }
-
-    public set_mouse_buttons(buttons: MouseButtons[]) {
-        this._mouse_buttons = buttons;
-    }
-
-    public set_keys(keys: string[]) {
-        this._keys = keys;
-    }
-
-    public set_modifiers(modifiers: KeyModifiers[]) {
-        this._modifiers = modifiers;
-    }
-}
-
-export class Keybind {
-    keybind_name: string;
-    maps: KeybindMap[];
-
-    constructor(name: string, maps: KeybindMap[]) {
-        this.keybind_name = name;
-        this.maps = maps;
-    }
-
-    public add_map(map: KeybindMap) {
-        if (this.maps.includes(map)) {
-            return;
-        }
-        this.maps = [...this.maps, map];
-    }
-
-    public remove_map(map: KeybindMap) {
-        this.maps = this.maps.filter((_map: KeybindMap) => {
-            return _map !== map;
-        })
-    }
-
-    public is_active(): boolean {
-        const active = this.maps.some(map => map.active)
-        return active;
-    }
-}
-
-export interface KeyEventData {
-    event: UIEvent
-}
-
-export interface EventData {
-    event?: UIEvent
-    node?: BaseNode,
-    slot?: NodeSlot
-}
-
-export class EventHandler {
-    name: string;
-    private func: (event_data: EventData) => void;
-
-    constructor(handler_name: string, handler_func: (event_data: EventData) => void) {
-        this.name = handler_name;
-        this.func = handler_func;
-    }
-
-    get handler() { return this.func; }
-}
-
-export interface HandlersInterface {
+export interface KeyHandlersInterface {
     just_activated?: (data: KeyEventData) => void;
     while_active?: (data: KeyEventData) => void;
     cleanup?: (data: KeyEventData) => void;
@@ -196,6 +65,7 @@ export class KeyEventManager implements ComponentEventHandler {
             this.event_handlers.set(key, []);
         });
     }
+   
     
 
     public set_event_handler(target_event: InputEvents, handler: EventHandler) {
@@ -206,7 +76,7 @@ export class KeyEventManager implements ComponentEventHandler {
         event_handlers.push(handler);
     }
 
-    public set_keybind_handler(keybind: Keybind, handlers: HandlersInterface) {
+    public set_keybind_handler(keybind: Keybind, handlers: KeyHandlersInterface) {
         this.keybinds.set(keybind, new KeyHandlers(
             handlers.just_activated != undefined ? handlers.just_activated : () => {},
             handlers.while_active != undefined ? handlers.while_active : () => {},
@@ -284,13 +154,6 @@ export class KeyEventManager implements ComponentEventHandler {
         handlers?.forEach(handler => {
             handler.handler({event: e});
         });
-        // this.update_modifier_states(e);
-
-        // this.update_other_states(e);
-
-        // this._set_other_states(Events.POINTER_MOVING, true)
-        // this.handle_keybinds({event: e})
-        // this._set_other_states(Events.POINTER_MOVING, false)
     }
 
     onPointerDown(e: PointerEvent): void {
@@ -326,9 +189,23 @@ export class KeyEventManager implements ComponentEventHandler {
     }
 
     onHoverNode(node: BaseNode): void {
-        
+        const handlers = this.event_handlers.get(InputEvents.HOVER_NODE)
+        handlers?.forEach(handler => {
+            handler.handler({node: node});
+        });
     }
+
     onHoverSlot(slot: NodeSlot): void {
-        
+        const handlers = this.event_handlers.get(InputEvents.HOVER_SLOT)
+        handlers?.forEach(handler => {
+            handler.handler({slot: slot});
+        });
+    }
+
+    onHoverBackground(): void {
+        const handlers = this.event_handlers.get(InputEvents.HOVER_BACKGROUND)
+        handlers?.forEach(handler => {
+            handler.handler({});
+        });
     }
 }
